+++ app/models/restaurant.model.js
@@ -1,5 +1,4 @@
 const mongoose = require("mongoose");
-
 const restaurantModel = new mongoose.Schema({
   name: String,
   address: String,
@@ -7,7 +6,6 @@ const restaurantModel = new mongoose.Schema({
   description: String,
   imgUrl: String
 });
-
 const Restaurant = mongoose.model('Restaurant', restaurantModel);
 
 module.exports = { Restaurant };
