+++ index.js
@@ -1,14 +1,12 @@
 require('dotenv').config();
 const express = require('express');
-const restaurantRoutes = require('./app/routes/restaurantRoutes');
+const restaurantRoutes = require('./app/routes/restaurant.routes');
 const { initDatabase } = require('./app/utils/initMongoDB');
 const bodyParser = require('body-parser');
 
 const app = express();
 
-console.log("before initDatabase");
 initDatabase();
-console.log("after initDatabase");
 
 app.use(bodyParser.json({ limit: '10mb' }));
 app.use(bodyParser.urlencoded({ limit: '10mb', extended: true }));
