+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "nodemon index.js"
   },
   "keywords": [],
   "author": "",
@@ -13,6 +13,8 @@
     "body-parser": "^1.20.2",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
+    "js-levenshtein": "^1.1.6",
+    "mongo-sanitize": "^1.1.0",
     "mongoose": "^8.2.3"
   }
 }
